prepare structure
