+++ pycrostates/segmentation/__init__.py
@@ -1,3 +1,12 @@
 from .segmentation import EpochsSegmentation, RawSegmentation  # noqa: F401
+from .transitions import (  # noqa: F401
+    compute_expected_transition_matrix,
+    compute_transition_matrix,
+)
 
-__all__ = ("RawSegmentation", "EpochsSegmentation")
+__all__ = (
+    "compute_expected_transition_matrix",
+    "compute_transition_matrix",
+    "EpochsSegmentation",
+    "RawSegmentation",
+)
